added static_cast to message preparation (undo implicit conversion to int)

+++ include/sha1.hpp
@@ -30,7 +30,7 @@ namespace libcrypt{
             std::uint32_t a, b, c, d, e, i, j, t;
 
             for(i = 0, j = 0; i < 16; ++i, j += 4)
-                m[i] = (data[j] << 24) + (data[j + 1] << 16) + (data[j + 2] << 8) + (data[j + 3]);
+                m[i] = static_cast<std::uint32_t>((data[j] << 24) + (data[j + 1] << 16) + (data[j + 2] << 8) + (data[j + 3]));
             for(; i < 80; ++i){
                 m[i] = (m[i - 3] ^ m[i - 8] ^ m[i - 14] ^ m[i - 16]);
                 m[i] = (m[i] << 1) | (m[i] >> 31);
